#100-011

+++ src/BankManagement.Domain/Models/ElasticSearchs/TransactionElasticModel.cs
@@ -9,7 +9,7 @@ public class TransactionElasticModel
     public Guid TransactionTypeId { get; set; }
     public string SenderIban { get; set; }
     public string ReceiverIban { get; set; }
-    public float Balance { get; set; }
+    public double Balance { get; set; }
     public DateTime CreationTime { get; set; }
     public DateTime? ElasticCreationTime { get; set; }
 }
